requirements.txt
Flask==3.1.2
Flask-Login==0.6.3
Flask-PyMongo==3.0.1
pymongo==4.14.1
bcrypt==4.3.0
python-dotenv==1.1.1
pytz==2025.2
gunicorn==23.0.0
resend==2.4.0
